+++ tests/test_console.py
@@ -768,3 +768,22 @@ def test_is_terminal_broken_file():
 def test_detect_color_system():
     console = Console(_environ={"TERM": "rxvt-unicode-256color"}, force_terminal=True)
     assert console._detect_color_system() == ColorSystem.EIGHT_BIT
+
+
+def test_reset_height():
+    """Test height is reset when rendering complex renderables."""
+    # https://github.com/Textualize/rich/issues/2042
+    class Panels:
+        def __rich_console__(self, console, options):
+            yield Panel("foo")
+            yield Panel("bar")
+
+    console = Console(force_terminal=False, width=20)
+
+    with console.capture() as capture:
+        console.print(Panel(Panels()), height=12)
+    result = capture.get()
+    print(repr(result))
+    expected = "╭──────────────────╮\n│ ╭──────────────╮ │\n│ │ foo          │ │\n│ ╰──────────────╯ │\n│ ╭──────────────╮ │\n│ │ bar          │ │\n│ ╰──────────────╯ │\n│                  │\n│                  │\n│                  │\n│                  │\n╰──────────────────╯\n"
+
+    assert result == expected